fix: rename sqlite3.py to sqlite3_utils.py and correct imports - avoid runtime errors
