+++ libsql_client/sqlite3_utils.py
